+++ operator/pkg/controllers/target_controllers.go
@@ -40,7 +40,7 @@ func (r *AIStoreReconciler) ensureTargetPrereqs(ctx context.Context, ais *aisv1.
 
 	// 2. Deploy services
 	svc := target.NewTargetHeadlessSvc(ais)
-	if _, err = r.k8sClient.CreateOrUpdateResource(ctx, ais, svc); err != nil {
+	if err = r.k8sClient.Apply(ctx, svc); err != nil {
 		r.recordError(ctx, ais, err, "Failed to deploy target SVC")
 		return
 	}
@@ -519,8 +519,7 @@ func (r *AIStoreReconciler) enableTargetExternalService(ctx context.Context,
 	targetSVCList := target.NewLoadBalancerSVCList(ais)
 	// 1. Try creating a LoadBalancer service for each target pod
 	for _, svc := range targetSVCList {
-		_, err := r.k8sClient.CreateOrUpdateResource(ctx, ais, svc)
-		if err != nil {
+		if err := r.k8sClient.Apply(ctx, svc); err != nil {
 			return err
 		}
 	}