+++ operator/pkg/resources/proxy/services.go
@@ -8,9 +8,10 @@ import (
 	aisapc "github.com/NVIDIA/aistore/api/apc"
 	aisv1 "github.com/ais-operator/api/v1beta1"
 	"github.com/ais-operator/pkg/resources/cmn"
+	"github.com/ais-operator/pkg/resources/ownerref"
 	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
+	corev1ac "k8s.io/client-go/applyconfigurations/core/v1"
 )
 
 const (
@@ -47,72 +48,60 @@ func ServiceSelectorLabels(aisName string) map[string]string {
 	}
 }
 
-// NewProxyHeadlessSvc returns a headless k8s services associated with `proxies`
-func NewProxyHeadlessSvc(ais *aisv1.AIStore) *corev1.Service {
+// NewProxyHeadlessSvc creates the apply config for the headless Service fronting proxy pods.
+func NewProxyHeadlessSvc(ais *aisv1.AIStore) *corev1ac.ServiceApplyConfiguration {
 	servicePort := ais.Spec.ProxySpec.ServicePort
 	controlPort := ais.Spec.ProxySpec.IntraControlPort
 	dataPort := ais.Spec.ProxySpec.IntraDataPort
 
-	return &corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      headlessSVCName(ais.Name),
-			Namespace: ais.Namespace,
-			Annotations: map[string]string{
-				"prometheus.io/scrape": "true",
-			},
-			Labels: cmn.NewServiceLabels(ais.Name, ServiceLabelHeadless),
-		},
-		Spec: corev1.ServiceSpec{
-			ClusterIP:                "None", // headless
-			PublishNotReadyAddresses: true,
-			Ports: []corev1.ServicePort{
-				{
-					Name:       "pub",
-					Protocol:   corev1.ProtocolTCP,
-					Port:       int32(servicePort.IntValue()),
-					TargetPort: servicePort,
-				},
-				{
-					Name:       "control",
-					Protocol:   corev1.ProtocolTCP,
-					Port:       int32(controlPort.IntValue()),
-					TargetPort: controlPort,
-				},
-				{
-					Name:       "data",
-					Protocol:   corev1.ProtocolTCP,
-					Port:       int32(dataPort.IntValue()),
-					TargetPort: dataPort,
-				},
-			},
-			Selector: ServiceSelectorLabels(ais.Name),
-		},
-	}
+	return corev1ac.Service(headlessSVCName(ais.Name), ais.Namespace).
+		WithOwnerReferences(ownerref.NewControllerRef(ais)).
+		WithAnnotations(map[string]string{
+			"prometheus.io/scrape": "true",
+		}).
+		WithLabels(cmn.NewServiceLabels(ais.Name, ServiceLabelHeadless)).
+		WithSpec(corev1ac.ServiceSpec().
+			WithClusterIP("None").
+			WithPublishNotReadyAddresses(true).
+			WithPorts(
+				corev1ac.ServicePort().
+					WithName("pub").
+					WithProtocol(corev1.ProtocolTCP).
+					WithPort(int32(servicePort.IntValue())).
+					WithTargetPort(servicePort),
+				corev1ac.ServicePort().
+					WithName("control").
+					WithProtocol(corev1.ProtocolTCP).
+					WithPort(int32(controlPort.IntValue())).
+					WithTargetPort(controlPort),
+				corev1ac.ServicePort().
+					WithName("data").
+					WithProtocol(corev1.ProtocolTCP).
+					WithPort(int32(dataPort.IntValue())).
+					WithTargetPort(dataPort),
+			).
+			WithSelector(ServiceSelectorLabels(ais.Name)),
+		)
 }
 
-func NewProxyLoadBalancerSVC(ais *aisv1.AIStore) *corev1.Service {
+func NewProxyLoadBalancerSVC(ais *aisv1.AIStore) *corev1ac.ServiceApplyConfiguration {
 	servicePort := ais.Spec.ProxySpec.ServicePort
 	publicNetPort := ais.Spec.ProxySpec.PublicPort
-	return &corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      loadBalancerSVCName(ais),
-			Namespace: ais.Namespace,
-			Annotations: map[string]string{
-				"prometheus.io/scrape": "true",
-			},
-			Labels: cmn.NewServiceLabels(ais.Name, ServiceLabelLB),
-		},
-		Spec: corev1.ServiceSpec{
-			Type: corev1.ServiceTypeLoadBalancer,
-			Ports: []corev1.ServicePort{
-				{
-					Name:       "pub",
-					Protocol:   corev1.ProtocolTCP,
-					Port:       int32(servicePort.IntValue()),
-					TargetPort: publicNetPort,
-				},
-			},
-			Selector: ServiceSelectorLabels(ais.Name),
-		},
-	}
+	return corev1ac.Service(loadBalancerSVCName(ais), ais.Namespace).
+		WithOwnerReferences(ownerref.NewControllerRef(ais)).
+		WithAnnotations(map[string]string{
+			"prometheus.io/scrape": "true",
+		}).
+		WithLabels(cmn.NewServiceLabels(ais.Name, ServiceLabelLB)).
+		WithSpec(corev1ac.ServiceSpec().
+			WithType(corev1.ServiceTypeLoadBalancer).
+			WithPorts(
+				corev1ac.ServicePort().
+					WithName("pub").
+					WithProtocol(corev1.ProtocolTCP).
+					WithPort(int32(servicePort.IntValue())).
+					WithTargetPort(publicNetPort),
+			).
+			WithSelector(ServiceSelectorLabels(ais.Name)),
+		)
 }