+++ operator/pkg/resources/target/services.go
@@ -10,9 +10,10 @@ import (
 	aisapc "github.com/NVIDIA/aistore/api/apc"
 	aisv1 "github.com/ais-operator/api/v1beta1"
 	"github.com/ais-operator/pkg/resources/cmn"
+	"github.com/ais-operator/pkg/resources/ownerref"
 	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
+	corev1ac "k8s.io/client-go/applyconfigurations/core/v1"
 )
 
 const (
@@ -53,85 +54,70 @@ func ServiceSelectorLabels(aisName string) map[string]string {
 	}
 }
 
-func NewTargetHeadlessSvc(ais *aisv1.AIStore) *corev1.Service {
+func NewTargetHeadlessSvc(ais *aisv1.AIStore) *corev1ac.ServiceApplyConfiguration {
 	servicePort := ais.Spec.TargetSpec.ServicePort
 	controlPort := ais.Spec.TargetSpec.IntraControlPort
 	dataPort := ais.Spec.TargetSpec.IntraDataPort
-	return &corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      headlessSVCName(ais.Name),
-			Namespace: ais.Namespace,
-			Annotations: map[string]string{
-				"prometheus.io/scrape": "true",
-			},
-			Labels: cmn.NewServiceLabels(ais.Name, ServiceLabelHeadless),
-		},
-		// TODO: in re G115 "integer overflow conversion int", see:
-		// https://github.com/kubernetes/apimachinery/blob/master/pkg/util/intstr/intstr.go#L32-L36
-		// once resolved in apimachinery, remove G115 exclusion from golangci yaml
-		Spec: corev1.ServiceSpec{
-			ClusterIP:                "None", // headless
-			PublishNotReadyAddresses: true,
-			Ports: []corev1.ServicePort{
-				{
-					Name:       "pub",
-					Protocol:   corev1.ProtocolTCP,
-					Port:       int32(servicePort.IntValue()),
-					TargetPort: servicePort,
-				},
-				{
-					Name:       "control",
-					Protocol:   corev1.ProtocolTCP,
-					Port:       int32(controlPort.IntValue()),
-					TargetPort: controlPort,
-				},
-				{
-					Name:       "data",
-					Protocol:   corev1.ProtocolTCP,
-					Port:       int32(dataPort.IntValue()),
-					TargetPort: dataPort,
-				},
-			},
-			Selector: ServiceSelectorLabels(ais.Name),
-		},
-	}
+	return corev1ac.Service(headlessSVCName(ais.Name), ais.Namespace).
+		WithOwnerReferences(ownerref.NewControllerRef(ais)).
+		WithAnnotations(map[string]string{
+			"prometheus.io/scrape": "true",
+		}).
+		WithLabels(cmn.NewServiceLabels(ais.Name, ServiceLabelHeadless)).
+		WithSpec(corev1ac.ServiceSpec().
+			WithClusterIP("None").
+			WithPublishNotReadyAddresses(true).
+			WithPorts(
+				corev1ac.ServicePort().
+					WithName("pub").
+					WithProtocol(corev1.ProtocolTCP).
+					WithPort(int32(servicePort.IntValue())).
+					WithTargetPort(servicePort),
+				corev1ac.ServicePort().
+					WithName("control").
+					WithProtocol(corev1.ProtocolTCP).
+					WithPort(int32(controlPort.IntValue())).
+					WithTargetPort(controlPort),
+				corev1ac.ServicePort().
+					WithName("data").
+					WithProtocol(corev1.ProtocolTCP).
+					WithPort(int32(dataPort.IntValue())).
+					WithTargetPort(dataPort),
+			).
+			WithSelector(ServiceSelectorLabels(ais.Name)),
+		)
 }
 
-func NewTargetLoadBalancerSVC(ais *aisv1.AIStore, targetIndex int32) *corev1.Service {
+func NewTargetLoadBalancerSVC(ais *aisv1.AIStore, targetIndex int32) *corev1ac.ServiceApplyConfiguration {
 	servicePort := ais.Spec.TargetSpec.ServicePort
 	publicNetPort := ais.Spec.TargetSpec.PublicPort
 	selectors := ServiceSelectorLabels(ais.Name)
 	selectors["statefulset.kubernetes.io/pod-name"] = fmt.Sprintf("%s-%d", statefulSetName(ais), targetIndex)
-	return &corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      loadBalancerSVCName(ais, targetIndex),
-			Namespace: ais.Namespace,
-			Annotations: map[string]string{
-				"prometheus.io/scrape": "true",
-			},
-			Labels: cmn.NewServiceLabels(ais.Name, ServiceLabelLB),
-		},
-		Spec: corev1.ServiceSpec{
-			Type: corev1.ServiceTypeLoadBalancer,
-			Ports: []corev1.ServicePort{
-				{
-					Name:       "pub",
-					Protocol:   corev1.ProtocolTCP,
-					Port:       int32(servicePort.IntValue()),
-					TargetPort: publicNetPort,
-				},
-			},
-			Selector: selectors,
-		},
-	}
+	return corev1ac.Service(loadBalancerSVCName(ais, targetIndex), ais.Namespace).
+		WithOwnerReferences(ownerref.NewControllerRef(ais)).
+		WithAnnotations(map[string]string{
+			"prometheus.io/scrape": "true",
+		}).
+		WithLabels(cmn.NewServiceLabels(ais.Name, ServiceLabelLB)).
+		WithSpec(corev1ac.ServiceSpec().
+			WithType(corev1.ServiceTypeLoadBalancer).
+			WithPorts(
+				corev1ac.ServicePort().
+					WithName("pub").
+					WithProtocol(corev1.ProtocolTCP).
+					WithPort(int32(servicePort.IntValue())).
+					WithTargetPort(publicNetPort),
+			).
+			WithSelector(selectors),
+		)
 }
 
-func NewLoadBalancerSVCList(ais *aisv1.AIStore) []*corev1.Service {
+func NewLoadBalancerSVCList(ais *aisv1.AIStore) []*corev1ac.ServiceApplyConfiguration {
 	return LoadBalancerSVCList(ais, 0, ais.GetTargetSize())
 }
 
-func LoadBalancerSVCList(ais *aisv1.AIStore, first, size int32) []*corev1.Service {
-	svcs := make([]*corev1.Service, 0, size)
+func LoadBalancerSVCList(ais *aisv1.AIStore, first, size int32) []*corev1ac.ServiceApplyConfiguration {
+	svcs := make([]*corev1ac.ServiceApplyConfiguration, 0, size)
 	for i := first; i < first+size; i++ {
 		svcs = append(svcs, NewTargetLoadBalancerSVC(ais, i))
 	}